+++ src/api/createSimpleSchema.ts
@@ -1,4 +1,5 @@
 import { Props, ModelSchema } from "./types"
+import object from "../types/object"
 
 /**
  * Creates a model schema that (de)serializes from / to plain javascript objects.
@@ -17,10 +18,10 @@ import { Props, ModelSchema } from "./types"
  * @returns model schema
  */
 export default function createSimpleSchema<T extends Object>(props: Props): ModelSchema<T> {
-    return {
+    return object({
         factory: function() {
             return {} as any
         },
         props: props
-    }
+    })
 }