+++ src/types/alias.ts
@@ -1,11 +1,6 @@
-import {
-    invariant,
-    isPropSchema,
-    isAliasedPropSchema,
-    isIdentifierPropSchema
-} from "../utils/utils"
+import { invariant, isSchema, isAliasedSchema, isIdentifierSchema } from "../utils/utils"
 import { _defaultPrimitiveProp } from "../constants"
-import { PropSchema, PropDef } from "../api/types"
+import { Schema, PropDef } from "../api/types"
 
 /**
  * Alias indicates that this model property should be named differently in the generated json.
@@ -21,16 +16,16 @@ import { PropSchema, PropDef } from "../api/types"
  * @param name name of the json field to be used for this property
  * @param propSchema propSchema to (de)serialize the contents of this field
  */
-export default function alias(name: string, propSchema?: PropDef): PropSchema {
+export default function alias(name: string, propSchema?: PropDef): Schema {
     invariant(name && typeof name === "string", "expected prop name as first argument")
     propSchema = !propSchema || propSchema === true ? _defaultPrimitiveProp : propSchema
-    invariant(isPropSchema(propSchema), "expected prop schema as second argument")
-    invariant(!isAliasedPropSchema(propSchema), "provided prop is already aliased")
+    invariant(isSchema(propSchema), "expected prop schema as second argument")
+    invariant(!isAliasedSchema(propSchema), "provided prop is already aliased")
     return {
         jsonname: name,
         serializer: propSchema.serializer,
         deserializer: propSchema.deserializer,
-        identifier: isIdentifierPropSchema(propSchema) || undefined,
+        identifier: isIdentifierSchema(propSchema) || undefined,
         beforeDeserialize: propSchema.beforeDeserialize,
         afterDeserialize: propSchema.afterDeserialize
     }