+++ src/types/identifier.ts
@@ -1,10 +1,6 @@
 import { invariant, processAdditionalPropArgs } from "../utils/utils"
 import { _defaultPrimitiveProp } from "../constants"
-import { AdditionalPropArgs, PropSchema, RegisterFunction } from "../api/types"
-
-const defaultRegisterFunction: RegisterFunction = (id, value, context) => {
-    context.rootContext.resolve(context.modelSchema, id, context.target)
-}
+import { AdditionalPropArgs, Schema, RegisterFunction } from "../api/types"
 
 /**
  *
@@ -43,8 +39,8 @@ const defaultRegisterFunction: RegisterFunction = (id, value, context) => {
 export function identifier(
     registerFn?: RegisterFunction,
     additionalArgs?: AdditionalPropArgs
-): PropSchema
-export function identifier(additionalArgs: AdditionalPropArgs): PropSchema
+): Schema
+export function identifier(additionalArgs: AdditionalPropArgs): Schema
 export default function identifier(
     arg1?: RegisterFunction | AdditionalPropArgs,
     arg2?: AdditionalPropArgs
@@ -61,20 +57,11 @@ export default function identifier(
         !additionalArgs || typeof additionalArgs === "object",
         "Additional property arguments should be an object, register function should be omitted or a funtion"
     )
-    let result: PropSchema = {
+    let result: Schema = {
         identifier: true,
+        // registerFn,
         serializer: _defaultPrimitiveProp.serializer,
-        deserializer: function(jsonValue, done, context) {
-            _defaultPrimitiveProp.deserializer(
-                jsonValue,
-                function(err, id) {
-                    defaultRegisterFunction(id, context.target, context)
-                    if (registerFn) registerFn(id, context.target, context)
-                    done(err, id)
-                },
-                context
-            )
-        }
+        deserializer: _defaultPrimitiveProp.deserializer
     }
     result = processAdditionalPropArgs(result, additionalArgs)
     return result