+++ src/types/list.ts
@@ -1,14 +1,14 @@
 import { SKIP } from "../constants"
 import {
     invariant,
-    isPropSchema,
-    isAliasedPropSchema,
+    isSchema,
+    isAliasedSchema,
     parallel,
     processAdditionalPropArgs
 } from "../utils/utils"
-import { onAfterDeserialize, onBeforeDeserialize } from "../core/deserialize"
 import { _defaultPrimitiveProp } from "../constants"
-import { AdditionalPropArgs, PropSchema } from "../api/types"
+import { AdditionalPropArgs, Schema } from "../api/types"
+import { doDeserialize } from "../core/deserialize"
 
 /**
  * List indicates that this property contains a list of things.
@@ -39,17 +39,11 @@ import { AdditionalPropArgs, PropSchema } from "../api/types"
  * @param propSchema to be used to (de)serialize the contents of the array
  * @param additionalArgs optional object that contains beforeDeserialize and/or afterDeserialize handlers
  */
-export default function list(
-    propSchema: PropSchema,
-    additionalArgs?: AdditionalPropArgs
-): PropSchema {
+export default function list(propSchema: Schema, additionalArgs?: AdditionalPropArgs): Schema {
     propSchema = propSchema || _defaultPrimitiveProp
-    invariant(isPropSchema(propSchema), "expected prop schema as first argument")
-    invariant(
-        !isAliasedPropSchema(propSchema),
-        "provided prop is aliased, please put aliases first"
-    )
-    let result: PropSchema = {
+    invariant(isSchema(propSchema), "expected prop schema as first argument")
+    invariant(!isAliasedSchema(propSchema), "provided prop is aliased, please put aliases first")
+    let result: Schema = {
         serializer: function(ar) {
             if (ar === undefined) {
                 return SKIP
@@ -60,59 +54,12 @@ export default function list(
         deserializer: function(jsonArray, done, context) {
             if (!Array.isArray(jsonArray)) return void done("[serializr] expected JSON array")
 
-            function processItem(
-                jsonValue: any,
-                onItemDone: (err?: any, value?: any) => void,
-                itemIndex: number
-            ) {
-                function callbackBefore(err: any, value: any) {
-                    if (!err) {
-                        propSchema.deserializer(value, deserializeDone, context)
-                    } else {
-                        onItemDone(err)
-                    }
-                }
-
-                function deserializeDone(err: any, value: any) {
-                    if (typeof propSchema.afterDeserialize === "function") {
-                        onAfterDeserialize(
-                            callbackAfter,
-                            err,
-                            value,
-                            jsonValue,
-                            jsonArray,
-                            itemIndex,
-                            context,
-                            propSchema
-                        )
-                    } else {
-                        onItemDone(err, value)
-                    }
-                }
-
-                function callbackAfter(errPreliminary: any, finalOrRetryValue: any) {
-                    if (
-                        errPreliminary &&
-                        finalOrRetryValue !== undefined &&
-                        typeof propSchema.afterDeserialize === "function"
-                    ) {
-                        propSchema.deserializer(finalOrRetryValue, deserializeDone, context)
-                    } else {
-                        onItemDone(errPreliminary, finalOrRetryValue)
-                    }
-                }
-
-                onBeforeDeserialize(
-                    callbackBefore,
-                    jsonValue,
-                    jsonArray,
-                    itemIndex,
-                    context,
-                    propSchema
-                )
-            }
-
-            parallel(jsonArray, processItem, done)
+            parallel(
+                jsonArray,
+                (jsonValue, onItemDone, itemIndex) =>
+                    doDeserialize(onItemDone, jsonValue, jsonArray, itemIndex, context, propSchema),
+                done
+            )
         }
     }
     result = processAdditionalPropArgs(result, additionalArgs)